Create group route, combine data fetching in layout
