+++ lxl-web/src/routes/(app)/[[lang=lang]]/(find)/SearchCard.svelte
