+++ lxl-web/src/routes/(app)/[[lang=lang]]/(find)/search.ts
